fix: Reorganize file structure
