+++ client/src/components/modals/Modal.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import CloseIcon from "./icons/CloseIcon.vue";
+import CloseIcon from "../icons/CloseIcon.vue";
 // Modal is gonna receive the props
 defineProps<{
   open: Boolean
