+++ client/src/components/modals/new-modals/NewProjectModal.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import Modal from './Modal.vue';
+import Modal from '../Modal.vue';
 </script>
 
 <template>
