+++ client/src/components/slots/MiniLabel.vue
